refactor: - simplify ErrorPrompt logic - rename file - make variant not optional - remove unnecessary useMemo form upgrade_license_prompt.tsx
